Refactor health information model to use integers for weight and height; update related UI components and remove unused files

+++ lib/features/registration_details/data/model/health_info_model.dart
@@ -1,6 +1,6 @@
 class PersonalInfoModel {
-  final double weight;
-  final double height;
+  final int weight;
+  final int height;
   final String dateOfBirth;
   final String gender;
   final List<String> foodAllergies;
